requirements.txt
absl-py==0.9.0
antlr4-python3-runtime==4.8
astor==0.8.1
attrs==19.3.0
decorator==4.4.1
gast==0.2.2
google-pasta==0.1.8
grpcio==1.27.1
h5py==2.10.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2
numpy==1.18.1
opt-einsum==3.1.0
Pillow==7.0.0
protobuf==3.11.3
psutil==5.6.7
PyYAML==5.3
scipy==1.4.1
six==1.14.0
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
topi==0.7.dev0
tornado==6.0.3
tvm==0.7.dev0
Werkzeug==1.0.0
wrapt==1.11.2
xgboost==0.90
